refactor: extract FlatRTree into shared header with KNN search
Move FlatRTree from anonymous namespace in spatial_join_physical.cpp
to a shared header (flat_rtree.hpp) so it can be reused by the
upcoming KNN join operator.

Changes:
- Extract FlatRTree class to src/spatial/operators/flat_rtree.hpp
- Add KNNSearch method using Hjaltason-Samet priority queue traversal
- Add FlatRTreeKNNState for KNN search state management
- Add Box2D::MinDistanceSquared for KNN distance lower bound
- Use Allocator instead of BufferManager (in-memory only)
- No functional changes to existing SPATIAL_JOIN behavior
